requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-cloud-bigquery>=3.0.0
functions-framework>=3.0.0
gunicorn>=20.1.0
